requirements.txt
fastapi          
uvicorn  
tensorflow     
numpy          
opencv-python-headless
Pillow            
python-multipart   
matplotlib          
